Use a ".sha" extension on the metafile
The previous method stored metadata for "./file.ext" as "./.shasummary/file.ext".
The problem is that this seems ripe to confuse some software into thinking that
the metafile is a file of type ".ext" implied by its extension.  Therefore, we
add a ".sha" to the metafile such that it would now be saved as
"./.shasummary/file.ext.sha" in an attempt to disambiguate

Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>